Render struct fields fields under struct code blocks
Add structs for fields in a type struct and render the fields additionally under the codeblock of the struct because we use large documentation for the fields so that the original presentation would be unreadable. Also modify the standard templates so that only the type struct will be rendered with the index.